Refactor: extract the logic for parsing projects into a seperate file :scissors:.
