+++ .idea/dictionaries/prodigy.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="prodigy">
+    <words>
+      <w>kitspace</w>
+    </words>
+  </dictionary>
+</component>
