+++ .idea/vcs.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="VcsDirectoryMappings">
+    <mapping directory="$PROJECT_DIR$" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/boards/github.com/8BitMixtape/8BitmixtapeNEO_ShenzhenReady" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/boards/github.com/kitspace-forks/Bus_Pirate" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/boards/github.com/kitspace-forks/arduino-uno" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/boards/github.com/kitspace-forks/hoverboard-breakout" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/boards/github.com/mattvenn/esp8266-breakout" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/boards/github.com/robotsrulz/SP_Adapter" vcs="Git" />
+  </component>
+</project>
